+++ tests/testthat/test-two_phase_sampling.R
@@ -18,18 +18,22 @@ test_that("two_phase_eif returns an uncentered EIF", {
 
   # assert that result is identical to two_phase_eif's
   expect_equal(
-    two_phase_eif(R = R,
-                  two_phase_weights = two_phase_weights,
-                  eif = eif,
-                  eif_predictions = eif_predictions,
-                  plugin_est = plugin_est),
+    two_phase_eif(
+      R = R,
+      two_phase_weights = two_phase_weights,
+      eif = eif,
+      eif_predictions = eif_predictions,
+      plugin_est = plugin_est
+    ),
     uncentered_eif
   )
 })
 
 
-context(paste("Estimators of nuisance parameters match manual analogs closely",
-              "in two-phase sampling designs"))
+context(paste(
+  "Estimators of nuisance parameters match manual analogs closely",
+  "in two-phase sampling designs"
+))
 source("utils_interventional.R")
 source("utils_natural.R")
 
@@ -71,9 +75,11 @@ u_learners <- v_learners <- hal_gaussian_lrnr
 data <- sim_medoutcon_data(n_obs = n_samp)
 w_names <- str_subset(colnames(data), "W")
 m_names <- str_subset(colnames(data), "M")
-data[, `:=`(R = rbinom(n_samp, 1, 0.9),
-            two_phase_weights = 1,
-            obs_weights = 1)]
+data[, `:=`(
+  R = rbinom(n_samp, 1, 0.9),
+  two_phase_weights = 1,
+  obs_weights = 1
+)]
 w <- as_tibble(data)[, w_names]
 a <- data$A
 z <- data$Z
@@ -193,8 +199,10 @@ test_that("MSE of pseudo-outcome used in v estimation is sufficiently low", {
 # Integration tests for estimators
 ################################################################################
 
-context(paste("Estimators of natural effects are close to true parameters in",
-              "two-phase sampling designs"))
+context(paste(
+  "Estimators of natural effects are close to true parameters in",
+  "two-phase sampling designs"
+))
 source("utils_natural.R")
 
 # packages
@@ -213,9 +221,11 @@ n_obs <- 500
 
 # 1) get data and column names for sl3 tasks (for convenience)
 data <- make_nide_data(n_obs = n_obs)
-data[, `:=`(R = rbinom(n_obs, 1, 0.9),
-            two_phase_weights = 1,
-            obs_weights = 1)]
+data[, `:=`(
+  R = rbinom(n_obs, 1, 0.9),
+  two_phase_weights = 1,
+  obs_weights = 1
+)]
 w_names <- str_subset(colnames(data), "W")
 m_names <- str_subset(colnames(data), "Z")
 
@@ -345,12 +355,14 @@ nie_true <- mean(EY_A1_Z1 - EY_A1_Z0)
 # 5) testing estimators for the NDE
 test_that("NDE: One-step estimate is near DGP truth", {
   expect_equal(nde_os$theta, nde_true,
-               tol = 1.96 * sqrt(var(nde_os$eif) / n_obs))
+    tol = 1.96 * sqrt(var(nde_os$eif) / n_obs)
+  )
 })
 
 test_that("NDE: TML estimate is near DGP truth", {
   expect_equal(nde_tmle$theta, nde_true,
-               tol = 1.96 * sqrt(var(nde_tmle$eif) / n_obs))
+    tol = 1.96 * sqrt(var(nde_tmle$eif) / n_obs)
+  )
 })
 
 test_that("NDE: Mean of estimated EIF is nearly zero for the one-step", {
@@ -365,12 +377,14 @@ test_that("NDE: Mean of estimated EIF is approximately solved for the TMLE", {
 # 6) testing estimators for the NIE
 test_that("NIE: One-step estimate is near DGP truth", {
   expect_equal(nie_os$theta, nie_true,
-               tol = 1.96 * sqrt(var(nie_os$eif) / n_obs))
+    tol = 1.96 * sqrt(var(nie_os$eif) / n_obs)
+  )
 })
 
 test_that("NIE: TML estimate is near DGP truth", {
   expect_equal(nie_tmle$theta, nie_true,
-               tol = 1.96 * sqrt(var(nie_tmle$eif) / n_obs))
+    tol = 1.96 * sqrt(var(nie_tmle$eif) / n_obs)
+  )
 })
 
 test_that("NIE: Mean of estimated EIF is nearly zero for the one-step", {